+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.1 (Unreleased)
 
+- feat: Add restart counter for task and mark failed [[GH-11](https://github.com/umbracle/eth2-validator/issues/11)]
 - feat: Add logical and cli support to stop (as destroy) a container [[GH-3](https://github.com/umbracle/eth2-validator/issues/3)]
 - feat: Add volume storage [[GH-1](https://github.com/umbracle/eth2-validator/issues/1)]
 - feat: Add github ci workflow [[GH-10](https://github.com/umbracle/eth2-validator/issues/10)]